confirm_otg_storage_access_text +1
I have not translated "OTG", I do not know what it means, and I do not know if the average user know what it is. But maybe those do that need to select the root directory? I use to look for acronyms on acronymfinder.com, they have 3 suggestions: On-The-Go (USB 2.0 specification), On the Ground (sales) and Off the Ground (gaming) - it does not seem to be what you mean? Maybe I might as well just omit the OTG-term and translate 'choose the root folder of the device'?